+++ sample/jest.config.default.js
@@ -4,7 +4,6 @@
 
 export default {
   // Recommended for native ES6 use (Aug-20):
-  testEnvironment: 'jest-environment-node',
   transform: {},
 
   // Default is 5000. None of our tests take that long; fail fast.